@@ -66,7 +66,6 @@
     "@typescript-eslint/eslint-plugin": "^4.13.0",
     "@typescript-eslint/parser": "^4.13.0",
     "cross-env": "^5.2.0",
-    "dotenv": "^16.3.1",
     "esbuild": "^0.19.2",
     "esbuild-node-tsc": "^2.0.5",
     "eslint": "^7.0.0",
@@ -92,8 +91,8 @@
     "postinstall": "~0.11.0",
     "prettier": "^2.2.1",
     "punycode": "^1.4.1",
-    "react": "^16.8.0 || ^17 || ^18",
-    "react-dom": "^16.8.0 || ^17 || ^18",
+    "react": "^16.8.0 || ^17 || ^18 || ^19",
+    "react-dom": "^16.8.0 || ^17 || ^18 || ^19",
     "rimraf": "^6.0.1",
     "semver": "^7.3.5",
     "ts-jest": "29.1.1",
@@ -136,14 +135,14 @@
     "@ethersproject/pbkdf2": "=5.6.0",
     "@ethersproject/constants": "=5.6.0",
     "browserify-sign": "4.2.2",
-    "mobx": "6.13.1",
-    "mobx-react-lite": "4.0.7",
-    "mobx-utils": "6.0.8",
-    "react": "18.3.1",
-    "tronweb": "5.3.2",
-    "axios": "1.7.2",
+    "mobx": "6.15.0",
+    "mobx-react-lite": "4.1.1",
+    "mobx-utils": "6.1.1",
+    "react": "19.1.1",
+    "tronweb": "6.1.1",
+    "axios": "1.13.2",
     "axios-extensions": "3.1.7",
-    "react-dom": "18.3.1",
+    "react-dom": "19.1.1",
     "web3": "1.7.4",
     "web3-utils": "1.7.4",
     "ws": "8.17.1",
@@ -161,8 +160,7 @@
     "is-core-module": "2.13.1",
     "error-ex": "1.3.2",
     "simple-swizzle": "0.2.2",
-    "has-ansi": "5.0.1",
-    "react-native-gesture-handler": "2.19.0"
+    "has-ansi": "5.0.1"
   },
   "browser": {
     "assert": false
