@@ -1,14 +1,14 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import "regenerator-runtime/runtime";
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import { StoreProvider } from "./stores";
 
 import { App } from "./app";
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root")!);
+root.render(
   <StoreProvider>
     <App />
-  </StoreProvider>,
-  document.getElementById("root")
+  </StoreProvider>
 );
